AGENTIC/pyproject.toml
[project]
name = "agentic"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "langchain>=1.2.8",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-ollama>=1.0.1",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.7",
    "langsmith>=0.6.8",
    "mcp>=1.26.0",
    "python-dotenv>=1.2.1",
]
